switched from double to float

+++ include/lvr2/io/IOUtils.hpp
@@ -50,7 +50,7 @@ namespace lvr2
  * @param registration          A matrix representing a registration (i.e. transformation) that
  * @return Eigen::Matrix4d      The transformed registration matrix
  */
-Eigen::Matrix4d transformRegistration(const Eigen::Matrix4d& transform, const Eigen::Matrix4d& registration);
+Eigen::Matrix4f transformRegistration(const Eigen::Matrix4f& transform, const Eigen::Matrix4f& registration);
 
 /**
  * @brief   Loads an Euler representation of from a pose file
@@ -269,7 +269,7 @@ Eigen::Matrix4d inverseTransform(const Eigen::Matrix4d& transform);
  * @param rotation  The rotation in radians
  * @return          The Matrix representation of the Pose
  */
-Eigen::Matrix4d poseToMatrix(const Eigen::Vector3d& position, const Eigen::Vector3d& rotation);
+Eigen::Matrix4f poseToMatrix(const Eigen::Vector3f& position, const Eigen::Vector3f& rotation);
 
 /**
  * @brief   Extracts the Pose from a Matrix
@@ -278,7 +278,7 @@ Eigen::Matrix4d poseToMatrix(const Eigen::Vector3d& position, const Eigen::Vecto
  * @param position  Output for the position of the Pose
  * @param rotation  Output for the rotation in radians
  */
-void matrixToPose(const Eigen::Matrix4d& mat, Eigen::Vector3d& position, Eigen::Vector3d& rotation);
+void matrixToPose(const Eigen::Matrix4f& mat, Eigen::Vector3f& position, Eigen::Vector3f& rotation);
 
 } // namespace lvr2
