+++ include/lvr2/registration/EigenSVDPointAlign.hpp
@@ -37,13 +37,13 @@
 #include <vector>
 
 #include <Eigen/Dense>
-using Eigen::Matrix4d;
-using Eigen::Vector3d;
+using Eigen::Matrix4f;
+using Eigen::Vector3f;
 
 namespace lvr2
 {
 
-using PointPairVector = std::vector<std::pair<const Vector3d*, const Vector3d*>>;
+using PointPairVector = std::vector<std::pair<const Vector3f*, const Vector3f*>>;
 
 class EigenSVDPointAlign
 {
@@ -51,9 +51,9 @@ public:
     EigenSVDPointAlign() {};
     double alignPoints(
         const PointPairVector& pairs,
-        const Vector3d& centroid_m,
-        const Vector3d& centroid_d,
-        Matrix4d& align);
+        const Vector3f& centroid_m,
+        const Vector3f& centroid_d,
+        Matrix4f& align);
 };
 
 } /* namespace lvr2 */
