+++ src/tools/lvr2_registration/Main.cpp
@@ -272,7 +272,7 @@ int main(int argc, char** argv)
         }
 
         file.replace_extension(pose_format);
-        Matrix4d pose = getTransformationFromPose(file);
+        Matrix4f pose = getTransformationFromPose(file).cast<float>();
 
         align.setScan(i, make_shared<Scan>(model->m_pointCloud, pose));
     }
